refactor: 프로젝트 조회 로직 리팩토링 (#283)
